+++ WingspanPrototype1/WingspanPrototype1/Controller/Members/DeleteMember.cs
@@ -13,22 +13,31 @@ namespace WingspanPrototype1.Controller.Members
             // Get databse 
             var database = DatabaseConnection.GetDatabase();
 
-            // Get collection
-            var collection = database.GetCollection<BsonDocument>("Members");
-
-            // Create filter
-            var filter = Builders<BsonDocument>.Filter.Eq("_id", id);
-
-            // Delete document 
-            try
+            if (database != null)
             {
-                collection.DeleteOne(filter);
-                return true;
+                // Get collection
+                var collection = database.GetCollection<BsonDocument>("Members");
+
+                // Create filter
+                var filter = Builders<BsonDocument>.Filter.Eq("_id", id);
+
+                // Delete document 
+                try
+                {
+                    collection.DeleteOne(filter);
+                    return true;
+                }
+                catch (Exception)
+                {
+                    return false;
+                }
             }
-            catch (Exception)
+            else
             {
                 return false;
             }
+
+            
 
         }
     }
